requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9.12
fvcore
pyyaml
numpy
pandas
matplotlib
seaborn
scikit-learn
grad-cam
gradio
tqdm
einops
